+++ LissandraFilesystem/src/Lissandra.h
@@ -47,7 +47,7 @@ void mainLissandra();
 void iniciarServidor();
 void setearValoresLissandra(t_config * archivoConfig);
 void insertKeysetter(t_keysetter * clave, char* tabla);
-void selectKey(char* tabla, uint receivedKey);
+void selectKey(char* tabla, uint16_t receivedKey);
 
 
 #endif /* LISSANDRA_H_ */
